# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anemo",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/permissions/catalog": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "Catalog",
        "operationId": "catalog_api_permissions_catalog_get",
        "parameters": [
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/permissions/summary": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "Summary",
        "description": "What an agent run may do: the global settings, or those of an agent profile.",
        "operationId": "summary_api_permissions_summary_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/permissions/preview": {
      "post": {
        "tags": [
          "permissions"
        ],
        "summary": "Preview",
        "description": "Summary for unsaved settings, so the settings page can show the effect live.",
        "operationId": "preview_api_permissions_preview_post",
        "parameters": [
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tools": {
      "get": {
        "tags": [
          "permissions"
        ],
        "summary": "Tools",
        "operationId": "tools_api_tools_get",
        "parameters": [
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolOut"
                  },
                  "title": "Response Tools Api Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "List Folder",
        "operationId": "list_folder_api_files_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 1000,
              "default": "",
              "title": "Path"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_api_profiles_get",
        "parameters": [
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfileOut"
                  },
                  "title": "Response List Profiles Api Profiles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_api_profiles_post",
        "parameters": [
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/profiles/{profile_id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_profiles__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_api_profiles__profile_id__put",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Delete Profile",
        "operationId": "delete_profile_api_profiles__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          },
          {
            "name": "aiw_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Aiw Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CatalogOut": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryOut"
            },
            "type": "array",
            "title": "Categories"
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/LevelOut"
            },
            "type": "array",
            "title": "Levels"
          }
        },
        "type": "object",
        "required": [
          "categories",
          "levels"
        ],
        "title": "CatalogOut"
      },
      "CategoryOut": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "default_level": {
            "type": "string",
            "enum": [
              "deny",
              "ask",
              "ask_dangerous",
              "workspace",
              "autonomous"
            ],
            "title": "Default Level"
          },
          "workspace_scoped": {
            "type": "boolean",
            "title": "Workspace Scoped"
          }
        },
        "type": "object",
        "required": [
          "capability",
          "label",
          "description",
          "default_level",
          "workspace_scoped"
        ],
        "title": "CategoryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LevelOut": {
        "properties": {
          "level": {
            "type": "string",
            "enum": [
              "deny",
              "ask",
              "ask_dangerous",
              "workspace",
              "autonomous"
            ],
            "title": "Level"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "level",
          "label"
        ],
        "title": "LevelOut"
      },
...
      "ListingOut": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryOut"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "path",
          "entries"
        ],
        "title": "ListingOut"
      },
...
      "PermissionSettings": {
        "properties": {
          "levels": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "deny",
                "ask",
                "ask_dangerous",
                "workspace",
                "autonomous"
              ]
            },
            "type": "object",
            "title": "Levels"
          },
          "ceiling": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "deny",
                "ask",
                "ask_dangerous",
                "workspace",
                "autonomous"
              ]
            },
            "type": "object",
            "title": "Ceiling"
          },
          "extra_rules": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "type": "array",
            "title": "Extra Rules"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          },
          "plan_review": {
            "type": "string",
            "enum": [
              "off",
              "always"
            ],
            "title": "Plan Review",
            "default": "off"
          }
        },
        "type": "object",
        "title": "PermissionSettings",
        "description": "Stored in app_settings[\"permissions\"] (security-sensitive section)."
      },
      "PermissionSummary": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SummaryItem"
            },
            "type": "array",
            "title": "Items"
          },
          "max_steps": {
            "type": "integer",
            "title": "Max Steps"
          },
          "max_tool_calls": {
            "type": "integer",
            "title": "Max Tool Calls"
          },
          "max_runtime_s": {
            "type": "integer",
            "title": "Max Runtime S"
          },
          "max_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cost Usd"
          },
          "plan_review": {
            "type": "string",
            "title": "Plan Review",
            "default": "off"
          }
        },
        "type": "object",
        "required": [
          "items",
          "max_steps",
          "max_tool_calls",
          "max_runtime_s"
        ],
        "title": "PermissionSummary"
      },
...
      "ProfileIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "instructions": {
            "type": "string",
            "maxLength": 50000,
            "title": "Instructions",
            "default": ""
          },
          "icon": {
            "type": "string",
            "maxLength": 40,
            "title": "Icon",
            "default": "bot"
          },
          "default_model_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model Id"
          },
          "permission_levels": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "deny",
                "ask",
                "ask_dangerous",
                "workspace",
                "autonomous"
              ]
            },
            "type": "object",
            "title": "Permission Levels"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Limits"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_review": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "off",
                  "always"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Review"
          },
          "skill_mode": {
            "type": "string",
            "enum": [
              "all",
              "selected",
              "none"
            ],
            "title": "Skill Mode",
            "default": "all"
          },
          "skill_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Skill Ids"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProfileIn"
      },
      "ProfileOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "instructions": {
            "type": "string",
            "title": "Instructions"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "default_model_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model Id"
          },
          "permission_levels": {
            "additionalProperties": {
              "type": "string",
              "enum": [
                "deny",
                "ask",
                "ask_dangerous",
                "workspace",
                "autonomous"
              ]
            },
            "type": "object",
            "title": "Permission Levels"
          },
          "limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Limits"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_review": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "off",
                  "always"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Review"
          },
          "skill_mode": {
            "type": "string",
            "enum": [
              "all",
              "selected",
              "none"
            ],
            "title": "Skill Mode"
          },
          "skill_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Skill Ids"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "instructions",
          "icon",
          "default_model_id",
          "permission_levels",
          "limits",
          "plan_review",
          "skill_mode",
          "skill_ids",
          "created_at",
          "updated_at"
        ],
        "title": "ProfileOut"
      },
...
      "ToolOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "capability": {
            "type": "string",
            "title": "Capability"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "capability"
        ],
        "title": "ToolOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}